[Imp] Main toolbar: Allow icon groups to be hidden as well. At least one icon group needs to be always visible, though (https://bugs.openmpt.org/view.php?id=1889).
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@23525 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Mainbar.h
@@ -19,7 +19,7 @@
 
 OPENMPT_NAMESPACE_BEGIN
 
-enum class MainToolBarItem : uint8;
+enum class MainToolBarItem : uint32;
 class CMainToolBar;
 
 class CStereoVU: public CStatic
